style: 컴포넌트 분리
FeedCard내부 컴포넌트 분리
