+++ src/components/Common/Feed/Card/FeedCard.module.scss
@@ -12,36 +12,14 @@
     align-items: center;
     justify-content: space-between;
     margin-bottom: 1.5rem;
-    .profile {
-      display: flex;
-      flex-direction: row;
-      align-items: center;
+    .more {
+      width: 3.2rem;
+      height: 3.2rem;
+      padding: 2rem;
+      border-radius: 50%;
+      background: url('/assets/img/icon-more.svg') center/ 3rem 3rem no-repeat;
       &:hover {
         opacity: 0.5;
-      }
-      .profileImg {
-        width: 4.2rem;
-        height: 4.2rem;
-        margin-right: 1rem;
-        border-radius: 50%;
-        border: 1px solid $color-c4c4c4;
-      }
-      .author {
-        max-width: 30rem;
-        .user {
-          font-size: 1.6rem;
-          font-weight: 700;
-          white-space: nowrap;
-          overflow: hidden;
-          text-overflow: ellipsis;
-        }
-        .account {
-          font-size: 1.2rem;
-          color: $color-767676;
-          white-space: nowrap;
-          overflow: hidden;
-          text-overflow: ellipsis;
-        }
       }
     }
   }
