+++ src/components/Common/Feed/MoreButton/FeedMoreButton.module.scss
@@ -1,0 +1,10 @@
+.more {
+  width: 3.2rem;
+  height: 3.2rem;
+  padding: 2rem;
+  border-radius: 50%;
+  background: url('/assets/img/icon-more.svg') center/ 3rem 3rem no-repeat;
+  &:hover {
+    opacity: 0.5;
+  }
+}
